product-service/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
psycopg2-binary==2.9.5
python-dotenv
redis==4.4.2
Werkzeug==2.2.2
PyJWT